+++ internal/search/commit/commit.go
@@ -9,7 +9,6 @@ import (
 	"github.com/opentracing/opentracing-go/log"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
-	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
@@ -18,7 +17,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	"github.com/sourcegraph/sourcegraph/internal/search/job"
-	"github.com/sourcegraph/sourcegraph/internal/search/limits"
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
 	searchrepos "github.com/sourcegraph/sourcegraph/internal/search/repos"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
@@ -32,9 +30,7 @@ type SearchJob struct {
 	Query                gitprotocol.Node
 	RepoOpts             search.RepoOptions
 	Diff                 bool
-	HasTimeFilter        bool
 	Limit                int
-	CodeMonitorID        *int64
 	IncludeModifiedFiles bool
 
 	// CodeMonitorSearchWrapper, if set, will wrap the commit search with extra logic specific to code monitors.
@@ -57,40 +53,10 @@ func (j *SearchJob) Run(ctx context.Context, clients job.RuntimeClients, stream 
 		return nil, err
 	}
 
-	opts := j.RepoOpts
-	if opts.Limit == 0 {
-		opts.Limit = reposLimit(j.HasTimeFilter)
-	}
-
-	resultType := "commit"
-	if j.Diff {
-		resultType = "diff"
-	}
-
-	var repoRevs []*search.RepositoryRevisions
-	repos := searchrepos.Resolver{DB: clients.DB, Opts: opts}
-	err = repos.Paginate(ctx, func(page *searchrepos.Resolved) error {
-		if repoRevs = page.RepoRevs; page.Next != nil {
-			return newReposLimitError(opts.Limit, j.HasTimeFilter, resultType)
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	bounded := goroutine.NewBounded(8)
-	for _, repoRev := range repoRevs {
-		repoRev := repoRev // we close over repoRev in onMatches
-		if ctx.Err() != nil {
-			// Don't keep spinning up goroutines if context has been canceled,
-			// but make sure we still clean up any running goroutines.
-			return nil, errors.Append(ctx.Err(), bounded.Wait())
-		}
-
+	searchRepoRev := func(repoRev *search.RepositoryRevisions) error {
 		// Skip the repo if no revisions were resolved for it
 		if len(repoRev.Revs) == 0 {
-			continue
+			return nil
 		}
 
 		args := &protocol.SearchRequest{
@@ -122,15 +88,29 @@ func (j *SearchJob) Run(ctx context.Context, clients job.RuntimeClients, stream 
 			return err
 		}
 
-		bounded.Go(func() error {
-			if j.CodeMonitorSearchWrapper != nil {
-				return j.CodeMonitorSearchWrapper(ctx, clients.DB, clients.Gitserver, args, repoRev.Repo.ID, doSearch)
-			}
-			return doSearch(args)
-		})
+		if j.CodeMonitorSearchWrapper != nil {
+			return j.CodeMonitorSearchWrapper(ctx, clients.DB, clients.Gitserver, args, repoRev.Repo.ID, doSearch)
+		}
+		return doSearch(args)
 	}
 
-	return nil, bounded.Wait()
+	bounded := goroutine.NewBounded(4)
+	defer func() { err = errors.Append(err, bounded.Wait()) }()
+
+	repos := searchrepos.Resolver{DB: clients.DB, Opts: j.RepoOpts}
+	return nil, repos.Paginate(ctx, func(page *searchrepos.Resolved) error {
+		for _, repoRev := range page.RepoRevs {
+			repoRev := repoRev
+			if ctx.Err() != nil {
+				// Don't keep spinning up goroutines if context has been canceled
+				return ctx.Err()
+			}
+			bounded.Go(func() error {
+				return searchRepoRev(repoRev)
+			})
+		}
+		return nil
+	})
 }
 
 func (j SearchJob) Name() string {
@@ -145,7 +125,6 @@ func (j *SearchJob) Tags() []log.Field {
 		trace.Stringer("query", j.Query),
 		trace.Stringer("repoOpts", &j.RepoOpts),
 		log.Bool("diff", j.Diff),
-		log.Bool("hasTimeFilter", j.HasTimeFilter),
 		log.Int("limit", j.Limit),
 		log.Bool("includeModifiedFiles", j.IncludeModifiedFiles),
 	}
@@ -377,37 +356,4 @@ func protocolMatchToCommitMatch(repo types.MinimalRepo, diff bool, in protocol.C
 		MessagePreview: messagePreview,
 		ModifiedFiles:  in.ModifiedFiles,
 	}
-}
-
-func newReposLimitError(limit int, hasTimeFilter bool, resultType string) error {
-	if hasTimeFilter {
-		return &TimeLimitError{ResultType: resultType, Max: limit}
-	}
-	return &RepoLimitError{ResultType: resultType, Max: limit}
-}
-
-func reposLimit(hasTimeFilter bool) int {
-	searchLimits := limits.SearchLimits(conf.Get())
-	if hasTimeFilter {
-		return searchLimits.CommitDiffWithTimeFilterMaxRepos
-	}
-	return searchLimits.CommitDiffMaxRepos
-}
-
-type DiffCommitError struct {
-	ResultType string
-	Max        int
-}
-
-type (
-	RepoLimitError DiffCommitError
-	TimeLimitError DiffCommitError
-)
-
-func (*RepoLimitError) Error() string {
-	return "repo limit error"
-}
-
-func (*TimeLimitError) Error() string {
-	return "time limit error"
 }